% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MongrelDB.R
\docType{package}
\name{MongrelDB-package}
\alias{MongrelDB}
\alias{MongrelDB-package}
\title{MongrelDB: Pure R Client for MongrelDB}
\description{
A pure R HTTP client for mongreldb-server. Talks JSON over the Kit transaction, query, and SQL endpoints, with a typed condition() error hierarchy and a native query builder. Depends on the 'curl' package for the HTTP transport.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/visorcraft/MongrelDB-R}
  \item \url{https://www.mongreldb.com}
  \item Report bugs at \url{https://github.com/visorcraft/MongrelDB-R/issues}
}

}
\author{
\strong{Maintainer}: Visorcraft \email{noreply@visorcraft.com}

Authors:
\itemize{
  \item Visorcraft \email{noreply@visorcraft.com}
}

}
