% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{mongreldb_connect}
\alias{mongreldb_connect}
\title{Connect to a running mongreldb-server daemon.}
\usage{
mongreldb_connect(url, token = NULL, username = NULL, password = NULL)
}
\arguments{
\item{url}{Base URL of the daemon, e.g. \code{"http://127.0.0.1:8453"}.}

\item{token}{Optional bearer token for \code{--auth-token} mode.}

\item{username}{Optional username for \code{--auth-users} HTTP Basic mode.}

\item{password}{Optional password for HTTP Basic mode.}
}
\value{
A \code{mongreldb_client} object to pass to the other functions.
}
\description{
Connect to a running mongreldb-server daemon.
}
