% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{mongreldb_count}
\alias{mongreldb_count}
\title{Row count for a table.}
\usage{
mongreldb_count(client, table)
}
\arguments{
\item{client}{A \code{mongreldb_client} from \code{\link[=mongreldb_connect]{mongreldb_connect()}}.}

\item{table}{Table name.}
}
\value{
Integer row count.
}
\description{
Row count for a table.
}
