% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{mongreldb_delete}
\alias{mongreldb_delete}
\title{Delete a row by its internal row id.}
\usage{
mongreldb_delete(client, table, row_id)
}
\arguments{
\item{client}{A \code{mongreldb_client} from \code{\link[=mongreldb_connect]{mongreldb_connect()}}.}

\item{table}{Table name.}

\item{row_id}{Integer row id.}
}
\description{
Delete a row by its internal row id.
}
