% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{mongreldb_delete_by_pk}
\alias{mongreldb_delete_by_pk}
\title{Delete a row by its primary key value.}
\usage{
mongreldb_delete_by_pk(client, table, pk)
}
\arguments{
\item{client}{A \code{mongreldb_client} from \code{\link[=mongreldb_connect]{mongreldb_connect()}}.}

\item{table}{Table name.}

\item{pk}{Primary key value.}
}
\description{
Delete a row by its primary key value.
}
