% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{mongreldb_drop_table}
\alias{mongreldb_drop_table}
\title{Drop a table by name.}
\usage{
mongreldb_drop_table(client, name)
}
\arguments{
\item{client}{A \code{mongreldb_client} from \code{\link[=mongreldb_connect]{mongreldb_connect()}}.}

\item{name}{Table name.}
}
\description{
Drop a table by name.
}
