% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{mongreldb_earliest_retained_epoch}
\alias{mongreldb_earliest_retained_epoch}
\title{Get the oldest epoch still queryable with \verb{AS OF EPOCH}.}
\usage{
mongreldb_earliest_retained_epoch(client)
}
\arguments{
\item{client}{A \code{mongreldb_client}.}
}
\value{
The oldest retained epoch.
}
\description{
Get the oldest epoch still queryable with \verb{AS OF EPOCH}.
}
