% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{mongreldb_history_retention_epochs}
\alias{mongreldb_history_retention_epochs}
\alias{mongreldb_history_retention}
\title{Get the current history-retention window.}
\usage{
mongreldb_history_retention_epochs(client)

mongreldb_history_retention(client)
}
\arguments{
\item{client}{A \code{mongreldb_client}.}
}
\value{
The current retention window in epochs.
}
\description{
Get the current history-retention window.
}
