% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{mongreldb_schema}
\alias{mongreldb_schema}
\title{Full schema catalog.}
\usage{
mongreldb_schema(client)
}
\arguments{
\item{client}{A \code{mongreldb_client} from \code{\link[=mongreldb_connect]{mongreldb_connect()}}.}
}
\value{
Named list of table descriptors.
}
\description{
Full schema catalog.
}
