% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{mongreldb_schema_for}
\alias{mongreldb_schema_for}
\title{Descriptor for a single table.}
\usage{
mongreldb_schema_for(client, table)
}
\arguments{
\item{client}{A \code{mongreldb_client} from \code{\link[=mongreldb_connect]{mongreldb_connect()}}.}

\item{table}{Table name.}
}
\description{
Descriptor for a single table.
}
