% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{mongreldb_set_history_retention_epochs}
\alias{mongreldb_set_history_retention_epochs}
\alias{mongreldb_set_history_retention}
\title{Set the history-retention window.}
\usage{
mongreldb_set_history_retention_epochs(client, epochs)

mongreldb_set_history_retention(client, epochs)
}
\arguments{
\item{client}{A \code{mongreldb_client}.}

\item{epochs}{New retention window in epochs.}
}
\value{
The updated retention response from the server.
}
\description{
Set the history-retention window.
}
