% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{mongreldb_tables}
\alias{mongreldb_tables}
\title{List all table names.}
\usage{
mongreldb_tables(client)
}
\arguments{
\item{client}{A \code{mongreldb_client}.}
}
\value{
Character vector of table names.
}
\description{
List all table names.
}
