% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{mongreldb_upsert}
\alias{mongreldb_upsert}
\title{Upsert (insert or update on PK conflict).}
\usage{
mongreldb_upsert(client, table, cells, update_cells = NULL)
}
\arguments{
\item{client}{A \code{mongreldb_client} from \code{\link[=mongreldb_connect]{mongreldb_connect()}}.}

\item{table}{Table name.}

\item{cells}{Named list mapping column id to value.}

\item{update_cells}{Optional named list of cells to update on conflict.}
}
\description{
Upsert (insert or update on PK conflict).
}
